pyproject.toml
[project]
name = "pers-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "motor[srv]>=3.7.0",
    "odmantic>=1.0.2",
    "pymongo[srv]>=4.11.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
